Remove evaluator_util and integrate functions to evaluator_main and engine_utils
